Fix misuse of Any in contravariant pos
Visitor's T parameter is used in contravariant position, so the least
specific type is not Any, it's Nothing.

This change is necessary because my change in RefChecks is banning some
class definitions, to allow for better and more useful GADT reasoning.
The definitions in Types (starting from Reader.Delegate) were
accidentally triggering this check, because of its use of Any in
contravariant position.